[spec] fix unrelated WebMock spec failure
Loosten the specificity of the WebMock request being stubbed here; the
stub message was generated by webmock and copy-pasted into Rspec, which
worked for Faraday 0.10. Faraday 0.11 generated different, even more
detailed headers.

But are we really testing the specifc headers of Faraday here? No. it
doesn't matter at all.

This commit removes the over-specificity of this stub_request call to
reduce these sorts of spurious failures.