fix: render_batch receives jobs in hash form.
All occurrences of RequestService's `render_batch` receive their jobs input
in hash form but were expecting it to be an array.
This caused the on_error callbacks to always silently get called with a
nil job.

Since the consistent form of arguments to the RequestService's methods is a hash,
the mock jobs in its test are also wrapped in a hash ( instead of an
array ).